feat: extract BCLB service helper functions and reporting to separate files
Move getUserLimits/isAgeVerified/isKYCCompliant/isSelfExcluded/getDailyStake/getWeeklyStake/getMonthlyStake/isInCoolingOffPeriod from BCLBService methods to helpers.go. Move GenerateComplianceReport/GetUserComplianceStatus/UpdateUserLimits/SetSelfExclusion/RemoveSelfExclusion/generateUserStats/generateFinancialStats to reporting.go. Remove EventBus/BCLBConfig/ComplianceCheck/ComplianceViolation/ViolationSeverity/UserLimits/ComplianceReport
