+++ internal/compliance/bclb/reporting.go
@@ -1,0 +1,187 @@
+package compliance
+
+import (
+	"context"
+	"time"
+
+	"github.com/shopspring/decimal"
+)
+
+// GenerateComplianceReport generates a comprehensive compliance report
+func (s *BCLBService) GenerateComplianceReport(ctx context.Context, period string) (*ComplianceReport, error) {
+	report := &ComplianceReport{
+		Period:      period,
+		GeneratedAt: time.Now(),
+	}
+
+	// In a real implementation, these would be calculated from actual data
+	report.TotalChecks = 10000
+	report.PassedChecks = 8500
+	report.FailedChecks = 1500
+
+	// Generate user statistics
+	report.UserStatistics = s.generateUserStats()
+
+	// Generate financial statistics
+	report.FinancialStats = s.generateFinancialStats()
+
+	// Generate violations (sample data)
+	report.Violations = []ComplianceViolation{
+		{
+			Type:        "AGE_VERIFICATION",
+			Description: "Users without age verification",
+			Severity:    ViolationSeverityCritical,
+			Action:      "BLOCK_BET",
+		},
+		{
+			Type:        "KYC_COMPLIANCE",
+			Description: "Users without KYC verification",
+			Severity:    ViolationSeverityHigh,
+			Action:      "RESTRICT_ACCESS",
+		},
+		{
+			Type:        "DAILY_LIMIT",
+			Description: "Users exceeding daily betting limits",
+			Severity:    ViolationSeverityMedium,
+			Action:      "COOLING_OFF",
+		},
+	}
+
+	// Generate recommendations
+	report.Recommendations = []string{
+		"Increase age verification measures",
+		"Implement real-time limit enforcement",
+		"Enhance AML monitoring for large transactions",
+		"Improve user education on responsible gaming",
+		"Strengthen KYC verification processes",
+	}
+
+	return report, nil
+}
+
+// generateUserStats generates user compliance statistics
+func (s *BCLBService) generateUserStats() UserComplianceStats {
+	// In a real implementation, these would be calculated from actual database queries
+	return UserComplianceStats{
+		TotalUsers:         10000,
+		VerifiedUsers:      8000,
+		AgeVerifiedUsers:   7500,
+		SelfExcludedUsers:  500,
+		LimitBreachedUsers: 200,
+		SuspiciousUsers:    50,
+		ComplianceRate:     0.85,
+	}
+}
+
+// generateFinancialStats generates financial compliance statistics
+func (s *BCLBService) generateFinancialStats() FinancialComplianceStats {
+	// In a real implementation, these would be calculated from actual database queries
+	return FinancialComplianceStats{
+		TotalStake:             decimal.NewFromFloat(1000000),
+		TotalWinnings:          decimal.NewFromFloat(450000),
+		TotalDeposits:          decimal.NewFromFloat(200000),
+		TotalWithdrawals:       decimal.NewFromFloat(150000),
+		TaxCollected:           decimal.NewFromFloat(50000),
+		AMLFlaggedTransactions: 25,
+		LargeTransactions:      100,
+	}
+}
+
+// GetUserComplianceStatus gets detailed compliance status for a specific user
+func (s *BCLBService) GetUserComplianceStatus(ctx context.Context, userID string) (*ComplianceCheck, error) {
+	check := &ComplianceCheck{
+		CheckTime: time.Now(),
+		CheckType: "USER_STATUS",
+		UserID:    userID,
+	}
+
+	// Check age verification
+	if !s.isAgeVerified(userID) {
+		check.Violations = append(check.Violations, ComplianceViolation{
+			Type:        "AGE_NOT_VERIFIED",
+			Description: "User age not verified",
+			Severity:    ViolationSeverityCritical,
+			Action:      "RESTRICT_ACCESS",
+		})
+	}
+
+	// Check KYC compliance
+	if !s.isKYCCompliant(userID) {
+		check.Violations = append(check.Violations, ComplianceViolation{
+			Type:        "KYC_NOT_COMPLIANT",
+			Description: "User KYC verification not completed",
+			Severity:    ViolationSeverityHigh,
+			Action:      "RESTRICT_ACCESS",
+		})
+	}
+
+	// Check self-exclusion
+	if s.isSelfExcluded(userID) {
+		check.Violations = append(check.Violations, ComplianceViolation{
+			Type:        "SELF_EXCLUDED",
+			Description: "User is self-excluded from betting",
+			Severity:    ViolationSeverityCritical,
+			Action:      "BLOCK_ACCESS",
+		})
+	}
+
+	// Check current limits
+	userLimits := s.getUserLimits(userID)
+	dailyStake := s.getDailyStake(userID)
+	weeklyStake := s.getWeeklyStake(userID)
+	monthlyStake := s.getMonthlyStake(userID)
+
+	// Add warnings for approaching limits
+	if dailyStake.GreaterThan(userLimits.DailyLimit.Mul(decimal.NewFromFloat(0.9))) {
+		check.Warnings = append(check.Warnings, "Approaching daily betting limit")
+	}
+
+	if weeklyStake.GreaterThan(userLimits.WeeklyLimit.Mul(decimal.NewFromFloat(0.9))) {
+		check.Warnings = append(check.Warnings, "Approaching weekly betting limit")
+	}
+
+	if monthlyStake.GreaterThan(userLimits.MonthlyLimit.Mul(decimal.NewFromFloat(0.9))) {
+		check.Warnings = append(check.Warnings, "Approaching monthly betting limit")
+	}
+
+	check.Passed = len(check.Violations) == 0
+
+	return check, nil
+}
+
+// UpdateUserLimits updates user-specific betting limits
+func (s *BCLBService) UpdateUserLimits(ctx context.Context, userID string, limits UserLimits) error {
+	// In a real implementation, this would update the database
+	// For now, just log the update
+	s.publishComplianceEvent("compliance.limits_updated", map[string]interface{}{
+		"user_id": userID,
+		"limits":  limits,
+	})
+
+	return nil
+}
+
+// SetSelfExclusion sets self-exclusion for a user
+func (s *BCLBService) SetSelfExclusion(ctx context.Context, userID string, duration time.Duration, reason string) error {
+	// In a real implementation, this would update the database
+	endDate := time.Now().Add(duration)
+
+	s.publishComplianceEvent("compliance.self_exclusion_set", map[string]interface{}{
+		"user_id":  userID,
+		"duration": duration,
+		"end_date": endDate,
+		"reason":   reason,
+	})
+
+	return nil
+}
+
+// RemoveSelfExclusion removes self-exclusion for a user
+func (s *BCLBService) RemoveSelfExclusion(ctx context.Context, userID string) error {
+	// In a real implementation, this would update the database
+	s.publishComplianceEvent("compliance.self_exclusion_removed", map[string]interface{}{
+		"user_id": userID,
+	})
+
+	return nil
+}
